final changes

+++ Routes/Pizzas.js
@@ -123,71 +123,76 @@ router.get('/:id', async (req, res) => {
 const update = (overwrite = false) => async (req, res) => {
 
 
-   try {
+  let user = await User.findById(req.user._id)
 
-
-     const pizza = await Pizza.findByIdAndUpdate(
- 
-
-       req.params.id,
- 
-
-       req.sanitizedBody,
- 
-
-       {
- 
-
-         new: true,
-        
-        
-         overwrite,
-        
-        
-         runValidators: true
- 
-       }
-     )
-     
-     
-     if (!pizza) throw new Error('Resource not found')
- 
-
-     res.send({
-
-
-      data: pizza
-
-
-    })
-
-    
-   } catch (err) {
-
-
-     console.log(err.message);
-
-
-     sendResourceNotFound(req, res)
   
-    }
+  if(user.isStaff)
+  {
+
+    const pizza = await Pizza.findByIdAndUpdate(
+ 
+
+      req.params.id,
+
+
+      req.sanitizedBody,
+
+
+      {
+
+
+        new: true,
+       
+       
+        overwrite,
+       
+       
+        runValidators: true
+
+      }
+    )
+    
+    
+    if (!pizza) throw new Error('Resource not found')
+
+
+    res.send({
+
+
+     data: pizza
+
+
+   })
+  }
+  else{
+
+    console.log(err.message);
+
+
+    sendResourceNotFound(req, res)
+
+  }
+
 
 }
  
 
-router.patch('/:id', sanitizeBody, update((overwrite = false)))
+router.patch('/:id',authenticate , sanitizeBody, update((overwrite = false)))
 
 
-router.put('/:id', sanitizeBody, update((overwrite = true)))
+router.put('/:id', authenticate,sanitizeBody, update((overwrite = true)))
 
 
-router.delete('/:id', async (req, res) => {
-
- 
-   try {
+router.delete('/:id',authenticate ,async (req, res) => {
 
 
-     const pizza = await Pizza.findByIdAndRemove(req.params.id)
+  let user = await User.findById(req.user._id)
+
+  
+  if(user.isStaff)
+  {
+
+    const pizza = await Pizza.findByIdAndRemove(req.params.id)
  
 
      if (!pizza) throw new Error('Resource not found')
@@ -200,15 +205,15 @@ router.delete('/:id', async (req, res) => {
      
     
     })
- 
 
-   } catch (err) {
- 
-    
+  }
+  else{
+
     sendResourceNotFound(req, res)
  
 
-   }
+  }
+
 
 })
